+++ game/src/vertices.h
@@ -4,7 +4,7 @@
 typedef struct {
     float x,y,z;
     int texIndex;
-    int lightValue;
+    float lightValue;
 } vertex_t;
 
 extern ivec3 directions[6];
